refactor: Rename namespace Converters to JsonConverters
Add Json as prefix to JsonConverter type name
