+++ src/Api.Rest.Dtos/Data/AttributeDto.cs
@@ -18,7 +18,7 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Data
 	using System.Xml;
 	using JetBrains.Annotations;
 	using Zeiss.PiWeb.Api.Rest.Dtos.Converter;
-	using Zeiss.PiWeb.Api.Rest.Dtos.Converters;
+	using Zeiss.PiWeb.Api.Rest.Dtos.JsonConverters;
 
 	#endregion
 
@@ -28,7 +28,7 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Data
 	/// <remarks>This class is immutable.</remarks>
 	[Serializable]
 	[Newtonsoft.Json.JsonConverter( typeof( AttributeConverter ) )]
-	[JsonConverter( typeof( AttributeJsonConverter ) )]
+	[JsonConverter( typeof( JsonAttributeConverter ) )]
 	public readonly struct AttributeDto : IEquatable<AttributeDto>
 	{
 		#region members
