+++ src/Api.Rest.Dtos/Data/CatalogAttributeDefinitionDto.cs
@@ -15,7 +15,7 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Data
 	using System;
 	using System.Text.Json.Serialization;
 	using Zeiss.PiWeb.Api.Rest.Dtos.Converter;
-	using Zeiss.PiWeb.Api.Rest.Dtos.Converters;
+	using Zeiss.PiWeb.Api.Rest.Dtos.JsonConverters;
 
 	#endregion
 
@@ -23,7 +23,7 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Data
 	/// Defines an entity's attribute which is based on a <see cref="Catalog"/>.
 	/// </summary>
 	[Newtonsoft.Json.JsonConverter( typeof( AttributeDefinitionConverter ) )]
-	[JsonConverter( typeof( CatalogAttributeDefinitionJsonConverter ) )]
+	[JsonConverter( typeof( JsonCatalogAttributeDefinitionConverter ) )]
 	public sealed class CatalogAttributeDefinitionDto : AbstractAttributeDefinitionDto
 	{
 		#region constructors
