+++ src/Api.Rest.Dtos/Data/PathInformationDto.cs
@@ -19,7 +19,7 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Data
 	using System.Text;
 	using System.Text.Json.Serialization;
 	using Zeiss.PiWeb.Api.Rest.Dtos.Converter;
-	using Zeiss.PiWeb.Api.Rest.Dtos.Converters;
+	using Zeiss.PiWeb.Api.Rest.Dtos.JsonConverters;
 
 	#endregion
 
@@ -33,7 +33,7 @@ namespace Zeiss.PiWeb.Api.Rest.Dtos.Data
 	/// * A characteristic path element can not have part elements as children
 	/// </summary>
 	[Newtonsoft.Json.JsonConverter( typeof( PathInformationConverter ) )]
-	[JsonConverter( typeof( PathInformationJsonConverter ) )]
+	[JsonConverter( typeof( JsonPathInformationConverter ) )]
 	public sealed class PathInformationDto : IFormattable, IEnumerable<PathElementDto>
 	{
 		#region members
